// Web/RMA-Web-Shell/projects/penscare/src/app/shared-penscare/models/penscare-lookups.ts
import { Lookup } from 'projects/shared-models-lib/src/lib/lookup/lookup';

export class PenscareLookups {
  genders?: Lookup[];
  languages?: Lookup[];
  provinces?: Lookup[];
  maritalStatus?: Lookup[];
  idTypes?: Lookup[];
  countries?: Lookup[];
  beneficiaryTypes?: Lookup[];
  titles?: Lookup[];
  communicationTypes?: Lookup[];
  benefitTypes?: Lookup[];
  populationGroups?: Lookup[];
  marriageTypes?: Lookup[];
  designationTypes?: Lookup[];
}